+++ GradeAntPlus_Prototype/archive/agentsX.py
@@ -84,7 +84,7 @@ def human_interaction_tool(tool_context: ToolContext, prompt: str) -> dict:
     # Allow the user to exit the loop early
     if human_input.lower() in ['done', 'exit', 'got it', 'thanks', 'skip']:
         logger.info("HumanTool: User indicated conversation is complete. Escalating to exit loop.")
-        print("🔚 User requested to end conversation.")
+        print("📚 User requested to end conversation.")
         tool_context.escalate = True  # This signals the parent LoopAgent to stop
         return {"human_response": human_input, "status": "completed"}
 
@@ -109,16 +109,35 @@ def text_to_speech_tool(tool_context: ToolContext, script_text: str) -> dict:
     """Converts a given text script into an MP3 audio file and saves it."""
     logger.info("TTS Tool: Starting audio conversion.")
     try:
-        # Use the session ID to create a unique filename
-        output_path = output_dir / f"audio_summary.mp3"
+        # Check if a target audio path is set in the session state
+        target_audio_path = tool_context.state.get("target_audio_path")
+        
+        if target_audio_path:
+            # Use the predefined path from main.py
+            output_path = Path(target_audio_path)
+            logger.info(f"TTS Tool: Using target path: {output_path}")
+        else:
+            # Fallback to default path (for backward compatibility)
+            output_path = output_dir / f"audio_summary.mp3"
+            logger.info(f"TTS Tool: Using fallback path: {output_path}")
+        
+        # Ensure the directory exists
+        output_path.parent.mkdir(parents=True, exist_ok=True)
 
         # Create the gTTS object and save the file
         tts = gTTS(text=script_text, lang='en', slow=False)
         tts.save(str(output_path))
         
         logger.info(f"TTS Tool: Audio file saved successfully to {output_path}")
-        # Return the path to be stored in the session state
-        return {"audio_file_path": str(output_path)}
+        
+        # Verify file was created and has content
+        if output_path.exists() and output_path.stat().st_size > 0:
+            logger.info(f"TTS Tool: Audio file verified - size: {output_path.stat().st_size} bytes")
+            return {"audio_file_path": str(output_path)}
+        else:
+            logger.error("TTS Tool: Audio file was not created or is empty")
+            return {"audio_file_path": None, "error": "Audio file creation failed"}
+            
     except Exception as e:
         logger.error(f"TTS Tool: Failed to generate audio file. Error: {e}")
         return {"audio_file_path": None, "error": str(e)}
@@ -128,6 +147,7 @@ def text_to_speech_tool(tool_context: ToolContext, script_text: str) -> dict:
 human_tool = FunctionTool(func=human_interaction_tool)
 exit_tool = FunctionTool(func=exit_loop_tool)
 tts_tool = FunctionTool(func=text_to_speech_tool) # NEW: Create the tool instance
+
 
 # ==============================================================================
 # SECTION 3: AGENT DEFINITIONS (TeachingAgent is updated)
@@ -186,7 +206,7 @@ audio_script_agent = LlmAgent(
 tts_agent = LlmAgent(
     model=MODEL,
     name="TtsAgent",
-    instruction="You have been given a script in the {audio_script} variable. Use the text_to_speech_tool to convert this script to an audio file.",
+    instruction="You have been given a script in the {audio_script} variable. Use the text_to_speech_tool to convert this script to an audio file. Call the tool with the script text as the parameter.",
     tools=[tts_tool],
     output_key="tts_output"
 )
@@ -194,7 +214,7 @@ tts_agent = LlmAgent(
 # ==============================================================================
 # SECTION 4: CUSTOM AGENTS FOR ORCHESTRATION
 # ==============================================================================
-# --- Agent 7: Feedback Triage Agent ---
+# --- Agent 4: Feedback Triage Agent (Corrected Logic) ---
 class FeedbackTriageAgent(BaseAgent):
     def __init__(self):
         super().__init__(name="FeedbackTriageAgent")
@@ -202,15 +222,14 @@ class FeedbackTriageAgent(BaseAgent):
     async def _run_async_impl(self, context: InvocationContext) -> AsyncGenerator[Event, None]:
         logger.info("TriageAgent: Analyzing knowledge response.")
         
-        # Print all input data keys
-        # print("\n📋 INPUT DATA:")
-        # input_keys = ["question_id", "student_answer", "reference_answer", "question_text"]
-        # for key in input_keys:
-        #     value = context.session.state.get(key, "NOT FOUND")
-        #     print(f"  {key}: {value}")
-        
-        # Get and display knowledge response
         knowledge_dict = context.session.state.get("knowledge_response")
+        
+        status = "needs_feedback"  # Default to needing feedback for safety.
+        if isinstance(knowledge_dict, dict):
+            if not knowledge_dict.get("needs_feedback", True): # can add more complex logic here
+                status = "correct"  # if the knowledge agent deems the student's answer correct
+
+        # Get and display knowledge response (for debugging)
         print("\n🧠 KNOWLEDGE ANALYSIS:")
         if isinstance(knowledge_dict, dict):
             schema_keys = ["is_correct", "needs_feedback", "misconceptions", "missing_steps", 
@@ -220,10 +239,6 @@ class FeedbackTriageAgent(BaseAgent):
                 print(f"  {key}: {value}")
         else:
             print("  ERROR: Knowledge response is not a dictionary")
-        
-        status = "needs_feedback"
-        if isinstance(knowledge_dict, dict) and knowledge_dict.get("is_correct", False):
-            status = "correct"
         
         context.session.state["feedback_status"] = status
         logger.info(f"TriageAgent: Feedback status set to '{status}'.")
@@ -303,3 +318,4 @@ summary_pipeline = SequentialAgent(
     ]
 )
 # ==============================================================================
+
