+++ GradeAntPlus_Prototype/code/utils.py
@@ -2,6 +2,8 @@ import json
 import logging
 import os
 from pathlib import Path
+from typing import Dict, Any
+from datetime import datetime
 
 # ==============================================================================
 # SECTION 1: LOGGER CONFIGURATION
@@ -87,6 +89,39 @@ def load_questions_from_file(filename: str) -> list:
     except Exception as e:
         logger.error(f"An unexpected error occurred while loading '{filename}': {e}")
         return []
+
+async def create_markdown_summary(summary_data: Dict[str, Any], output_path: Path) -> None:
+    """Creating readable markdown summary"""
+    try:
+        markdown_content = f"""# GradeAnt+ Session Summary
+
+## Session Overview
+{summary_data.get('session_overview', 'No overview available')}
+
+## Key Concepts and Formulas
+{summary_data.get('key_concepts_and_formulas', 'No concepts listed')}
+
+## Common Misconceptions Addressed
+{summary_data.get('common_misconceptions_addressed', 'No misconceptions noted')}
+
+## Areas for Further Study
+{summary_data.get('areas_for_further_study', 'No recommendations available')}
+
+## Overall Performance Rating
+{summary_data.get('overall_performance_rating', 'Not rated')}
+
+---
+*Generated by GradeAnt+ on {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}*
+"""
+        
+        with open(output_path, 'w', encoding='utf-8') as f:
+            f.write(markdown_content)
+        
+        logger.info(f"Markdown summary created: {output_path}")
+        
+    except Exception as e:
+        logger.error(f"Failed to create markdown summary: {str(e)}")
+
 
 if __name__ == "__main__":
     INPUT_FOLDER = Path(__file__).parent.parent / "data" / "input"
