+++ GradeAntPlus_Prototype/data/output/SessionReport_qa_1a2fada6.json
@@ -1,0 +1,86 @@
+{
+  "session_summary": {
+    "session_overview": "This session covered basic kinematics (uniform acceleration) and basic capacitor principles (charge, voltage, and capacitance). The student demonstrated understanding of the Q=CV relationship but initially struggled with the units of acceleration.",
+    "key_concepts_and_formulas": "Key concepts covered were: 1. Acceleration (a = (v - u)/t), including its definition as the rate of change of velocity. 2. Units of measurement for acceleration (m/s²). 3. Capacitance, Charge, and Voltage relationship (Q = C * V).",
+    "common_misconceptions_addressed": "The primary misconception addressed was using incorrect units for acceleration (m/s^3 instead of m/s^2).",
+    "areas_for_further_study": "1. Review the definitions and units for fundamental kinematic quantities (displacement, velocity, and acceleration). Pay close attention to how these units are derived. 2. Practice unit conversions and dimensional analysis to ensure correct units in calculations."
+  },
+  "audio_summary_path": null,
+  "detailed_transcript": [
+    {
+      "question_data": {
+        "question_id": "q1",
+        "question_text": "A car accelerates uniformly from rest to 20 m/s in 5 seconds. What is its acceleration?",
+        "student_answer": "a = (v - u)/t = (20 - 0)/5 = 4 m/s^3",
+        "reference_answer": "Acceleration is given by a = (v - u)/t = (20 - 0)/5 = 4 m/s^2"
+      },
+      "knowledge_response": {
+        "is_correct": false,
+        "needs_feedback": true,
+        "misconceptions": [
+          "Incorrect units for acceleration (m/s^3 instead of m/s^2)."
+        ],
+        "missing_steps": [
+          "Stating the correct units for acceleration."
+        ],
+        "key_concepts": [
+          "Acceleration",
+          "Units of measurement"
+        ],
+        "hint": "What are the standard units for acceleration?",
+        "confidence": "high"
+      },
+      "conversation_history": [
+        {
+          "role": "agent",
+          "content": "You mentioned the units for acceleration are m/s^3. Can you explain how you arrived at that? What does acceleration actually measure?"
+        },
+        {
+          "role": "user",
+          "content": ""
+        },
+        {
+          "role": "agent",
+          "content": "Let's break it down. Acceleration is the rate of change of *velocity*. Velocity is measured in meters per second (m/s). So, if acceleration is how velocity changes over time, how would you express that mathematically, and what units would that imply?"
+        },
+        {
+          "role": "user",
+          "content": ""
+        },
+        {
+          "role": "agent",
+          "content": "Think about it this way: acceleration is the change in velocity (m/s) divided by time (s). So, what do you get when you divide m/s by s?  The correct units for acceleration are meters per second squared (m/s²). Does that make sense?"
+        },
+        {
+          "role": "user",
+          "content": ""
+        }
+      ],
+      "teaching_response": null
+    },
+    {
+      "question_data": {
+        "question_id": "q8",
+        "question_text": "A capacitor in a circuit has a capacitance of 2 µF and a voltage of 50 V. What is the charge stored?",
+        "student_answer": "Q = C * V = 2 * 50 = 100 µC",
+        "reference_answer": "Charge in a capacitor is Q = C * V = 2 µF * 50 V = 100 µC"
+      },
+      "knowledge_response": {
+        "is_correct": true,
+        "needs_feedback": false,
+        "misconceptions": [],
+        "missing_steps": [],
+        "key_concepts": [
+          "Capacitance",
+          "Charge",
+          "Voltage",
+          "Q=CV"
+        ],
+        "hint": "The student correctly calculated the charge. No hint is needed.",
+        "confidence": "high"
+      },
+      "conversation_history": [],
+      "teaching_response": null
+    }
+  ]
+}
