+++ GradeAntPlus_Prototype/data/output/SessionSummary_qa_1a2fada6.md
@@ -1,0 +1,19 @@
+# GradeAnt+ Session Summary
+
+## Session Overview
+This session covered basic kinematics (uniform acceleration) and basic capacitor principles (charge, voltage, and capacitance). The student demonstrated understanding of the Q=CV relationship but initially struggled with the units of acceleration.
+
+## Key Concepts and Formulas
+Key concepts covered were: 1. Acceleration (a = (v - u)/t), including its definition as the rate of change of velocity. 2. Units of measurement for acceleration (m/s²). 3. Capacitance, Charge, and Voltage relationship (Q = C * V).
+
+## Common Misconceptions Addressed
+The primary misconception addressed was using incorrect units for acceleration (m/s^3 instead of m/s^2).
+
+## Areas for Further Study
+1. Review the definitions and units for fundamental kinematic quantities (displacement, velocity, and acceleration). Pay close attention to how these units are derived. 2. Practice unit conversions and dimensional analysis to ensure correct units in calculations.
+
+## Overall Performance Rating
+Not rated
+
+---
+*Generated by GradeAnt+ on 2025-09-20 15:23:58*
