--- GradeAntPlus_Prototype/data/output/session_report_qa_e95a1c86.json
@@ -1,82 +1,0 @@
-{
-  "session_summary": {
-    "session_overview": "This session covered problems related to kinematics (specifically uniform acceleration) and capacitance. The student demonstrated understanding of the capacitance problem but initially struggled with the units in the acceleration problem, requiring a hint to correct their answer.",
-    "key_concepts_and_formulas": "Key concepts covered include: Acceleration, Capacitance, Voltage, Charge, Units of measurement. Formulas used: a = (v - u)/t, Q = C * V",
-    "common_misconceptions_addressed": "The primary misconception addressed was the incorrect use of units for acceleration (m/s^3 instead of m/s^2). This indicated a misunderstanding of what acceleration physically represents.",
-    "areas_for_further_study": "1. Review the definitions and units of kinematic quantities, paying close attention to the physical meaning of each unit. 2. Practice unit conversions and dimensional analysis to ensure correct units are used in calculations. 3. Work through additional problems involving uniformly accelerated motion, focusing on interpreting the results in terms of changing velocity."
-  },
-  "audio_summary_path": null,
-  "detailed_transcript": [
-    {
-      "question_data": {
-        "question_id": "q1",
-        "question_text": "A car accelerates uniformly from rest to 20 m/s in 5 seconds. What is its acceleration?",
-        "student_answer": "a = (v - u)/t = (20 - 0)/5 = 4 m/s^3",
-        "reference_answer": "Acceleration is given by a = (v - u)/t = (20 - 0)/5 = 4 m/s^2"
-      },
-      "knowledge_response": {
-        "is_correct": false,
-        "needs_feedback": true,
-        "misconceptions": [
-          "Incorrect units for acceleration. Acceleration is measured in m/s^2, not m/s^3."
-        ],
-        "missing_steps": [],
-        "key_concepts": [
-          "Acceleration",
-          "Units of measurement"
-        ],
-        "hint": "Check the units of acceleration. What does m/s^2 physically represent?",
-        "confidence": "high"
-      },
-      "conversation_history": [
-        {
-          "role": "agent",
-          "content": "The hint suggests thinking about what m/s^2 physically represents. Can you explain in words what 'meters per second squared' means in terms of how velocity changes over time?"
-        },
-        {
-          "role": "user",
-          "content": "thanks"
-        },
-        {
-          "role": "agent",
-          "content": "Let's break it down. If acceleration is measured in m/s^2, and we can think of it as (m/s)/s, what does the numerator (m/s) represent, and what does the denominator (s) represent? How does one affect the other?"
-        },
-        {
-          "role": "user",
-          "content": "thanks"
-        },
-        {
-          "role": "agent",
-          "content": "Acceleration is the rate of change of velocity. m/s^2 means that for every second, the velocity changes by a certain number of meters per second. So, if an object has an acceleration of 5 m/s^2, its velocity increases by 5 m/s every second. Do you understand this explanation?"
-        },
-        {
-          "role": "user",
-          "content": "thnks"
-        }
-      ]
-    },
-    {
-      "question_data": {
-        "question_id": "q8",
-        "question_text": "A capacitor in a circuit has a capacitance of 2 µF and a voltage of 50 V. What is the charge stored?",
-        "student_answer": "Q = C * V = 2 * 50 = 100 µC",
-        "reference_answer": "Charge in a capacitor is Q = C * V = 2 µF * 50 V = 100 µC"
-      },
-      "knowledge_response": {
-        "is_correct": true,
-        "needs_feedback": false,
-        "misconceptions": [],
-        "missing_steps": [],
-        "key_concepts": [
-          "Capacitance",
-          "Voltage",
-          "Charge",
-          "Q=CV"
-        ],
-        "hint": "The student's answer is correct. No hints are necessary.",
-        "confidence": "high"
-      },
-      "conversation_history": []
-    }
-  ]
-}
